Move v011 package contents to main package
- Move MachineBuilder.kt, StateMachine.kt, and Effect.kt from v011 to main package
- Update imports in all affected files
- Fix Mermaid diagram generation to match expected format
- Remove v011 package
