+++ lib/src/main/kotlin/app/cash/kfsm/Effect.kt
@@ -1,7 +1,4 @@
-package app.cash.kfsm.v011
-
-import app.cash.kfsm.State
-import app.cash.kfsm.Value
+package app.cash.kfsm
 
 fun interface Effect<ID, V : Value<ID, V, S>, S : State<ID, V, S>> {
   fun apply(v: V): Result<V>
